✨ feat: 4단계 완료 모달 연결 구현
